+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Breaking: In CMake, PajladaSerialize and PajladaSignals are now private so you need to include them yourself. (#164)
+- Dev: Don't export target under `Pajlada::` namespace. (#164)
+
 ## v0.4.0
 
 - Breaking: Removed `SettingManager`s `pp`, `gPP`, and `stringify` functions. (#153)